Read Bib file

Read a bib file into a data.frame

read_bib(file, skip = 0L, max_lines = NULL, encoding = "UTF-8")

Arguments

file
File or connection

skip
The lines to skip

max_lines
The maximum number of lines to read

encoding
Assumed encoding of file (passed to readLines)

Value

A data.frame with each row as a bib entry and each column as a field

Details

Inspired and partially credited to bib2df::bib2df() although this has no dependencies outside of base functions and much quicker. This speed seems to come from removing stringr functions and simplifying a few *apply functions. This will also include as many categories as possible from the entry.

See also

Examples

file <- "https://raw.githubusercontent.com/jmbarbone/bib-references/master/references.bib"
bibdf <- read_bib(file, max_lines = 51L)

if (package_available("tibble")) {
  tibble::as_tibble(bibdf)
} else {
  head(bibdf)
}
#> # A tibble: 4 × 13
#>   key    field  author journal title year  issn  month number pages volume doi  
#>   <chr>  <chr>  <chr>  <chr>   <chr> <chr> <chr> <chr> <chr>  <chr> <chr>  <chr>
#> 1 ames2… artic… Ames,… Journa… The … 2006  0092… aug   4      440-… 40     10.1…
#> 2 ander… artic… Ander… Psycho… Effe… 2001  NA    NA    5      353-… 12     10.1…
#> 3 ayduk… artic… Ayduk… Journa… Regu… 2000  NA    NA    5      776   79     10.1…
#> 4 baker… artic… Baker… Nature… 1,50… 2016  NA    NA    7604   452   533    10.1…
#> # … with 1 more variable: publisher <chr>

if (package_available("bib2df") & package_available("bench")) {
  file <- system.file("extdata", "bib2df_testfile_3.bib", package = "bib2df")

  # Doesn't include the 'tidying' up
  foo <- function(file) {
    bib <- ("bib2df" %colons% "bib2df_read")(file)
    ("bib2df" %colons% "bib2df_gather")(bib)
  }

# \donttest{
  bench::mark(
    `read_bib` = read_bib(file),
    `bib2df` = bib2df::bib2df(file),
    `foo` = foo(file),
    check = FALSE
  )[1:9]
# }
}
#> Warning: `as_data_frame()` was deprecated in tibble 2.0.0.
#> Please use `as_tibble()` instead.
#> The signature and semantics have changed, see `?as_tibble`.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.
#> # A tibble: 3 × 9
#>   expression     min  median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time
#>   <bch:expr> <bch:t> <bch:t>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm>
#> 1 read_bib     2.5ms  2.61ms     382.     9.39KB     4.46   171     2      448ms
#> 2 bib2df      13.4ms 13.76ms      72.4    2.64MB     4.67    31     2      428ms
#> 3 foo         4.76ms  4.93ms     202.    85.26KB     4.54    89     2      441ms